Stuart23 GitHub actions (#5)
Running the tests with `coverage` on TravisCI was giving the following error:

_WARNING: Testing via this command is deprecated and will be removed in a future version. Users looking for a generic test entry point independent of test runner are encouraged to use tox._

Instead, the tests are done with `tox`. I have switched the CI to GitHub actions from Travis. This is personal preference, but Actions is better integrated on GitHub than Travis. Also, if people are forking this into a private fork, they will still be able to utilize CI without having to sign up for a Travis subscription.